+++ sequential/lu_blocked.cc
@@ -13,7 +13,7 @@ using std::chrono::milliseconds;
 using std::chrono::seconds;
 using std::chrono::system_clock;
 
-void print_matrix(float *A, int N, int n)
+void print_matrix(double *A, int N, int n)
 {
   for (int i = 0; i < N; ++i)
   {
@@ -46,8 +46,8 @@ int main(int argc, char **argv)
   if ((N % B) != 0)
     n = B * (N / B) + B;
 
-  float *A = (float *)malloc(n * n * sizeof(float));
-  float *L = (float *)malloc(N * N * sizeof(float));
+  double *A = (double *)malloc(n * n * sizeof(double));
+  double *L = (double *)malloc(N * N * sizeof(double));
   for (int i = 0; i < N; ++i)
   {
     for (int j = 0; j < N; ++j)
@@ -111,12 +111,12 @@ int main(int argc, char **argv)
   {
     for (int j = 0; j < N; ++j)
     {
-      out_file.write((char *)&A[i * n + j], sizeof(float));
+      out_file.write((char *)&A[i * n + j], sizeof(double));
     }
   }
   for (int i = 0; i < N * N; ++i)
   {
-    out_file.write((char *)&L[i], sizeof(float));
+    out_file.write((char *)&L[i], sizeof(double));
   }
   out_file.close();
   free(A);
